requirements.txt
# Core packages
requests>=2.28.1
beautifulsoup4>=4.11.1
feedparser>=6.0.10
PyPDF2>=3.0.1
openai>=1.6.0
Flask>=2.2.2
SQLAlchemy>=1.4.46
pandas>=1.5.2
matplotlib>=3.6.2
gunicorn>=20.1.0
psutil>=5.9.0

# Web related
Werkzeug>=2.2.2
Jinja2>=3.1.2
Flask-WTF>=1.1.1
wtforms>=3.0.1

# Data processing
numpy>=1.23.5
scikit-learn>=1.2.0

# Database drivers 
# Uncomment the one you need depending on your database
# psycopg2-binary>=2.9.5  # For PostgreSQL
# pymysql>=1.0.2  # For MySQL
# pyodbc>=4.0.35  # For SQL Server

# Utilities
tqdm>=4.64.1
python-dotenv>=0.21.0 